Move event types into events dir.
